requirements.txt
discord.py==2.4.0
dotenv==0.0.5
colorama==0.4.6
tabulate==0.9.0